This is my first submission of this project to 'git.' Very unfamiliar with this process in general. Having said that, I will elaborate on project details.
Dipp is a project that I started in order to teach myself a little about
python and pygtk. It's going to be a RSS Aggregator and Twitter
client with the ability for add-on's and themes. The code that has been written so far has been done so with very little programming experience and a 'learn as we go mentality'

The code is somewhat commented throughout, and there is a "lot" well.. a
"shit-load" of bad code that needs to be cleaned up or even scrapped.

Let's have fun with this and see what we can learn.

Thanks, solotitan